requirements.txt
streamlit==1.26.0
tensorflow==2.13.0
numpy==1.26.2
pandas==2.1.1
matplotlib==3.8.1
Pillow==10.1.0